Remove Section ‘upcoming events’ image from homepage
The conference is over. Packing up. Going home. Boy was that a great time
in the ole Bay Area. I’m so glad we had our event there. Couldn’t have
thought it go any better - but sure enought, it was great times to be 
had by all. 

If you ever need a good reason to go there, try the flank steak on the 
fisherman’s wharf. The corn I had there was the best in my life. It was
wrapped in some kind of coffee seasoning - probably just coffee beans 
now that I think about it!

The only other time I was in SF was on a road trip with my bros AP, S&S, 
& P. The 5 of us traveled around the country in an extended bed, extended cab, 
diesel ram truck. It was amazing. I think the whole trip cost me $600